+++ valkey_glide_commands_common.h
@@ -207,6 +207,11 @@ int execute_flushdb_command(zval* object, int argc, zval* return_value, zend_cla
 int execute_flushall_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
 int execute_bgsave_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
 int execute_bgrewriteaof_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
+int execute_client_pause_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
+int execute_client_unpause_command(zval*             object,
+                                   int               argc,
+                                   zval*             return_value,
+                                   zend_class_entry* ce);
 int execute_time_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
 int execute_scan_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
 int execute_cluster_scan_command(const void* glide_client,
@@ -719,6 +724,34 @@ int execute_unlink_command(zval* object, int argc, zval* return_value, zend_clas
         }                                                                               \
         zval_dtor(return_value);                                                        \
         RETURN_FALSE;                                                                   \
+    }
+
+#define CLIENT_PAUSE_METHOD_IMPL(class_name)                                            \
+    PHP_METHOD(class_name, clientPause) {                                               \
+        if (execute_client_pause_command(getThis(),                                     \
+                                         ZEND_NUM_ARGS(),                               \
+                                         return_value,                                  \
+                                         strcmp(#class_name, "ValkeyGlideCluster") == 0 \
+                                             ? get_valkey_glide_cluster_ce()            \
+                                             : get_valkey_glide_ce())) {                \
+            return;                                                                     \
+        }                                                                               \
+        zval_dtor(return_value);                                                        \
+        RETURN_FALSE;                                                                   \
+    }
+
+#define CLIENT_UNPAUSE_METHOD_IMPL(class_name)                                            \
+    PHP_METHOD(class_name, clientUnpause) {                                               \
+        if (execute_client_unpause_command(getThis(),                                     \
+                                           ZEND_NUM_ARGS(),                               \
+                                           return_value,                                  \
+                                           strcmp(#class_name, "ValkeyGlideCluster") == 0 \
+                                               ? get_valkey_glide_cluster_ce()            \
+                                               : get_valkey_glide_ce())) {                \
+            return;                                                                       \
+        }                                                                                 \
+        zval_dtor(return_value);                                                          \
+        RETURN_FALSE;                                                                     \
     }
 
 #define TIME_METHOD_IMPL(class_name)                                            \